Add web support for file picker
- filepicker popup launches filepicker depending on platform.
- web platform uses FilePickerCross.
- other platforms uses old "SystemFilePicker.
- separated getSaveFilesDirecory method into separate io supplier class in order to reduce depedencies.
