+++ kanbansim/lib/common/input_output_file_picker/input/filepicker_interface.dart
@@ -1,0 +1,10 @@
+import 'package:flutter/material.dart';
+
+abstract class FilePicker {
+  FilePicker({
+    @required this.returnPickedFilePath,
+  });
+
+  final Function(String) returnPickedFilePath;
+  void pickSaveFile();
+}
